# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ghost Protocol API",
    "version": "1.0.0",
    "description": "Machine-readable API surface for GhostRank discovery and GhostGate access."
  },
  "servers": [
    {
      "url": "https://ghostprotocol.cc"
    }
  ],
  "paths": {
...
    "/api/telemetry/x402/settlements": {
      "post": {
        "summary": "Persist merchant-signed x402 settlement evidence",
        "operationId": "reportX402Settlement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid settlement report"
          },
          "404": {
            "description": "Gateway config not found"
          },
          "409": {
            "description": "Gateway/service mismatch"
          },
          "429": {
            "description": "Rate limited"
          },
          "503": {
            "description": "x402 schema unavailable"
          }
        }
      }
    }
  }
}